Refactored initialize_models into models.py
